+++ iiif_pipeline/manifests.py
